attention: rephrasing the problem
Move the "why this specific approach" section to _before_ the "breaking
down the problem" section. Now we're building towards it in a more
straightforward way.

For now, I'm keeping the old text in a TODO. But I'll probably want to
turn it into a quick recap after the "breaking down".